Created crate enemy and sprite
